(big bada-boom) Initial commit
Yes. It is a rather monstrous commit for the very first one but I have a reason for this. I pulled this code out of another Rust app I am working on, refactored some things and made it a separate project. I still have some source to "recreate" in this project so the next couple commits might also be large.